cloud/requirements.txt
google-cloud-documentai>=2.20.0
google-cloud-storage>=2.10.0
